invert SchemaIO/SmartForm implementation
