+++ app/packages/components/src/components/SmartForm/RJSF/templates/ObjectFieldTemplate.tsx
