+++ app/packages/components/src/components/SmartForm/RJSF/templates/index.tsx
