+++ app/packages/components/src/components/SmartForm/RJSF/translators/ui.ts
