+++ app/packages/components/src/components/SmartForm/index.tsx
@@ -1,77 +1,38 @@
 import React from "react";
-import Form from "@rjsf/mui";
-import validator from "@rjsf/validator-ajv8";
-
-import {
-  convertRJSFDataToSchemaIO,
-  translateSchemaComplete,
-} from "./translators";
-
-import widgets from "./widgets";
-import templates from "./templates";
-
-export { isSchemaIOSchema, isJSONSchema } from "./translators";
-export type { SmartFormProps };
-
-import type { IChangeEvent } from "@rjsf/core";
-import type { ValidatorType, UiSchema } from "@rjsf/utils";
+import { SchemaIOComponent } from "@fiftyone/core/src/plugins/SchemaIO";
+import RJSF, { type RJSFProps } from "./RJSF";
 import type { SchemaType } from "@fiftyone/core/src/plugins/SchemaIO/utils/types";
+import type { RJSFSchema } from "@rjsf/utils";
+
+export { isSchemaIOSchema, isJSONSchema } from "./RJSF/translators";
 
 export interface SmartFormProps {
-  schema: SchemaType;
+  schema: SchemaType | RJSFSchema;
   data?: unknown;
-  uiSchema?: UiSchema;
-  validator?: ValidatorType;
-  onChange?: (data: unknown) => void;
+  id?: string;
+  useSchemaIO?: boolean;
+  onChange?: (data: unknown, liteValues?: Record<string, unknown>) => void;
+
+  // SchemaIO only
+  shouldClearUseKeyStores?: boolean;
+  onPathChange?: (
+    path: string,
+    value: unknown,
+    schema?: SchemaType,
+    updatedState?: unknown,
+    liteValue?: unknown
+  ) => void;
+
+  // RJSF only
   onSubmit?: (data: unknown) => void;
+  uiSchema?: RJSFProps["uiSchema"];
+  validator?: RJSFProps["validator"];
 }
 
 export default function SmartForm(props: SmartFormProps) {
-  const {
-    schema: jsonSchema,
-    uiSchema: generatedUiSchema,
-    warnings,
-    formData,
-  } = translateSchemaComplete(props.schema, props.data);
-
-  const mergedUiSchema = { ...generatedUiSchema, ...props.uiSchema };
-
-  if (warnings.length > 0) {
-    console.warn("[SmartForm] Schema translation warnings:", warnings);
-  }
-
-  const handleChange = (event: IChangeEvent, _id?: string) => {
-    if (props.onChange) {
-      const schemaIOData = convertRJSFDataToSchemaIO(
-        event.formData,
-        props.schema
-      );
-
-      props.onChange(schemaIOData);
-    }
-  };
-
-  const handleSubmit = (event: IChangeEvent, _nativeEvent: React.FormEvent) => {
-    if (props.onSubmit) {
-      const schemaIOData = convertRJSFDataToSchemaIO(
-        event.formData,
-        props.schema
-      );
-
-      props.onSubmit(schemaIOData);
-    }
-  };
-
-  return (
-    <Form
-      schema={jsonSchema}
-      uiSchema={mergedUiSchema}
-      validator={props.validator || validator}
-      widgets={widgets}
-      templates={templates}
-      formData={formData}
-      onChange={handleChange}
-      onSubmit={handleSubmit}
-    />
+  return props.useSchemaIO ? (
+    <SchemaIOComponent {...props} />
+  ) : (
+    <RJSF {...props} schema={props.schema as SchemaType} />
   );
 }
