+++ app/packages/core/src/plugins/SchemaIO/index.tsx
@@ -1,38 +1,90 @@
-import { SchemaIOComponent as SchemaIO } from "./SchemaIOComponent";
-import SmartForm, {
-  isJSONSchema,
-  type SmartFormProps,
-} from "../../../../components/src/components/SmartForm";
-import type { SchemaType } from "./utils/types";
-import type { RJSFSchema } from "@rjsf/utils";
+import { PluginComponentType, registerComponent } from "@fiftyone/plugins";
+import { cloneDeep, get, set } from "lodash";
+import { useCallback, useEffect, useRef } from "react";
+import DynamicIO from "./components/DynamicIO";
+import { SchemaIOContext, clearUseKeyStores } from "./hooks";
+import { coerceValue, getLiteValue } from "./utils";
 
-export interface SchemaIOComponentProps {
-  schema: SchemaType | RJSFSchema;
-  data?: unknown;
-  id?: string;
-  useJSONSchema?: boolean;
-  onChange?: (data: unknown, liteValues?: Record<string, unknown>) => void;
+export function SchemaIOComponent(props) {
+  const { onChange, onPathChange, id, shouldClearUseKeyStores, data } = props;
+  const stateRef = useRef({});
+  const autoFocused = useRef(false);
+  const schemaIOContext = { id };
+  const storeRef = useRef({ liteValues: {} });
 
-  // SchemaIO only
-  shouldClearUseKeyStores?: boolean;
-  onPathChange?: (
-    path: string,
-    value: unknown,
-    schema?: SchemaType,
-    updatedState?: unknown,
-    liteValue?: unknown
-  ) => void;
+  useEffect(() => {
+    if (data) {
+      stateRef.current = { ...data };
+    }
+  }, [data]);
 
-  // RJSF only
-  onSubmit?: (data: unknown) => void;
-  uiSchema?: SmartFormProps["uiSchema"];
-  validator?: SmartFormProps["validator"];
+  useEffect(() => {
+    return () => {
+      if (shouldClearUseKeyStores !== false) {
+        clearUseKeyStores(id);
+      }
+    };
+  }, []);
+
+  const onIOChange = useCallback(
+    (path, value, schema, ancestors) => {
+      const computedValue = coerceValue(path, value, schema);
+      const currentState = stateRef.current;
+      const updatedState = cloneDeep(currentState);
+      set(updatedState, path, cloneDeep(computedValue));
+      stateRef.current = updatedState;
+
+      const liteValue = getLiteValue(value, schema);
+      const store = storeRef.current;
+      if (liteValue) {
+        store.liteValues[path] = liteValue;
+      }
+
+      if (onPathChange) {
+        onPathChange(path, computedValue, schema, updatedState, liteValue);
+      }
+
+      if (onChange) {
+        onChange(updatedState, store.liteValues);
+      }
+
+      // propagate the change to all ancestors
+      for (const ancestorPath in ancestors) {
+        const ancestorSchema = ancestors[ancestorPath];
+        const ancestorValue = get(updatedState, ancestorPath);
+        if (onPathChange) {
+          onPathChange(
+            ancestorPath,
+            ancestorValue,
+            ancestorSchema,
+            updatedState
+          );
+        }
+      }
+
+      return updatedState;
+    },
+    [onChange, onPathChange]
+  );
+
+  return (
+    <SchemaIOContext.Provider value={schemaIOContext}>
+      <DynamicIO
+        {...props}
+        data={data}
+        root_id={id}
+        onChange={onIOChange}
+        path=""
+        autoFocused={autoFocused}
+      />
+    </SchemaIOContext.Provider>
+  );
 }
 
-export function SchemaIOComponent(props: SchemaIOComponentProps) {
-  if (props.useJSONSchema || isJSONSchema(props.schema)) {
-    return <SmartForm {...props} schema={props.schema as SchemaType} />;
-  } else {
-    return <SchemaIO {...props} />;
-  }
-}
+registerComponent({
+  name: "SchemaIOComponent",
+  label: "SchemaIOComponent",
+  component: SchemaIOComponent,
+  type: PluginComponentType.Component,
+  activator: () => true,
+});
